+++ lib/src/inherited_theme_provider.dart
@@ -3,11 +3,12 @@ import 'package:flutter/widgets.dart';
 /// Provides current mode to its children.
 /// This is an internal widget and should not be used directly.
 class InheritedThemeProvider extends InheritedWidget {
-  /// current mode
-  final int mode;
+  /// current theme index
+  final int index;
 
   // ignore: public_member_api_docs
-  const InheritedThemeProvider(this.mode, {super.key, required super.child});
+  const InheritedThemeProvider(
+      {super.key, required this.index, required super.child});
 
   /// Allows access to current mode.
   static InheritedThemeProvider? of(BuildContext context) {
@@ -16,6 +17,6 @@ class InheritedThemeProvider extends InheritedWidget {
 
   @override
   bool updateShouldNotify(covariant InheritedThemeProvider oldWidget) {
-    return mode != oldWidget.mode;
+    return index != oldWidget.index;
   }
 }